backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.26.8
botocore==1.29.165
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.5
django-cors-headers==3.13.0
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
drf-spectacular-sidecar==2023.9.1
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
ndg-httpsclient==0.5.1
Pillow==9.2.0
psycopg2-binary==2.9.5
pusher==3.3.2
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.2
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
